requirements.txt
asgiref==3.2.3
astroid==2.3.3
Django==3.0.3
django-crispy-forms==1.8.1
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
PyLin==0.2.7
pylint==2.4.4
pylint-django==2.0.13
pylint-plugin-utils==0.6
pyserial==3.4
pytz==2019.3
six==1.14.0
sqlparse==0.3.0
typed-ast==1.4.1
wrapt==1.11.2
